Early draft of converting to/from JSON objects.
Yeah, I'm not really clear on what to do with the information once I have
it. Most of them I can set up to create a new
object(Route/User/Session/etc.), but I'm not clear on the extent to which
I should be interacting with the database itself, or whether I'm just
going to be handing this data off to a yet-to-be-done receiver.